+++ src/sys/arch/namespace/instructions/portable.rs
@@ -3,7 +3,7 @@
 //!
 //
 
-use crate::{Arch, cfg_if};
+use crate::Arch;
 
 /// # Portable abstractions over architecture-dependent instructions.
 #[cfg(any_target_arch_linux)]
@@ -29,18 +29,12 @@ impl Arch {
     /// - The value is only meaningful for measuring relative durations on the same core.
     #[inline(always)]
     pub fn cycles() -> u64 {
-        cfg_if! {
-            if #[cfg(any(target_arch = "x86", target_arch = "x86_64"))] {
-                Arch::rdtsc()
-            } else if #[cfg(target_arch = "arm")] {
-                Arch::cntvct()
-            } else if #[cfg(target_arch = "aarch64")] {
-                Arch::cntvct()
-            } else if #[cfg(any(target_arch = "riscv32", target_arch = "riscv64"))] {
-                Arch::rdcycle().into()
-            } else {
-                compile_error!("Cycle counter not implemented for this architecture");
-            }
+        cfg_select! {
+                 any(target_arch = "x86", target_arch = "x86_64") => Arch::rdtsc(),
+                                              target_arch = "arm" => Arch::cntvct(),
+                                          target_arch = "aarch64" => Arch::cntvct(),
+            any(target_arch = "riscv32", target_arch = "riscv64") => Arch::rdcycle().into(),
+            _ => compile_error!("Cycle counter not implemented for this architecture"),
         }
     }
 }