+++ src/sys/io/mod.rs
@@ -32,10 +32,9 @@ crate::structural_mods! { // _mods
         #[cfg(any(feature = "std", all(not(feature = "std"), feature = "io")))]
         pub use super::io::*;
 
-        crate::cfg_if! { if #[cfg(feature = "std")] {
-            pub use super::std_reexports::*;
-        } else {
-            pub use super::no_std_define::*;
-        }}
+        cfg_select! {
+            feature = "std" => { pub use super::std_reexports::*; }
+                          _ => { pub use super::no_std_define::*; }
+        }
     }
 }